frost-signer/coordinator message signing (#48)
* store DkgPublicShares and use them to calculate aggregate public key

* add threshold field to Coordinator; store public nonces during sign op; exit loop when we have threshold nonces

* check for aggregate public key before trying to sign

* create an array of PolyCommitments and use it to make a SignatureAggregator

* add comments for rest of sign; add field for selected signer ids in SignatureShareRequest

* build SignatureShareRequest

* request signature shares

* receive signature shares from expected signers

* finish sign_message

* log agg sig on success; start db

* use current_dkg_id with nonce/sign messages; add DkgSign command so we don't need a db to test

* fix frost-signer tests to fill in added fields

* signer_id as u32; listen for SignShareRequest

* signer_id is u32

* reply with a SignatureShareResponse

* sig share array for each signer; u32 id

* respond to NonceRequest with PublicNonce

* display the msg to be signed

* display nonce activity

* change NonceRespose to per-party basis because controller doesnt know party_ids per signer

* shows whats going on with nonce gathering

* a single party per NonceResponse and SignatureShareRequest

* adapt from signer_id to party_id; use BTreeMap instead of HashMap to get ordered keys() output

* better output

* use total_parties for frost calls

* more output. remove party from wait queue on receipt of sig share

* update Cargo.lock

* wtfrost updates; PARTY_MAX is u32

* wtfrost update for frost-coordinator

* gather nonces from all parties; sign request needs a Vec<(party_id, PublicNonce)>; party sigs now good but group sig is still bad

* fix party and group sign

* remove old p256k1 dep in frost-coordinator; update Cargo.lock

* return group signature from sign_message

* signature request takes a Vec<(u32,PublicNonce)>; fmt fixes

---------

+++ frost-signer/src/config.rs
@@ -20,7 +20,7 @@ pub struct Common {
 // on-disk format for frost save data
 #[derive(Clone, Deserialize, Default, Debug)]
 pub struct Signer {
-    pub frost_id: u64,
+    pub frost_id: u32,
     pub frost_state_file: String,
 }
 
@@ -37,7 +37,7 @@ pub struct Cli {
 
     /// Turn debugging information on
     #[arg(short, long)]
-    id: Option<u64>,
+    id: Option<u32>,
 }
 
 impl Config {
